Towards #812: carve out functionality from reward_calculator.py
