+++ df_py/predictoor/test/test_predictoor_calc_rewards.py
@@ -10,7 +10,7 @@ from df_py.predictoor.calc_rewards import (
 )
 from df_py.predictoor.models import Prediction, Predictoor, PredictoorBase
 from df_py.util.networkutil import DEV_CHAINID
-from df_py.volume.reward_calculator import RewardShaper
+from df_py.util.reward_shaper import RewardShaper
 
 
 @pytest.fixture(autouse=True)
